follwows

+++ src/main/java/de/unidue/inf/is/user_profileServlet.java
@@ -195,7 +195,7 @@ public final class user_profileServlet extends HttpServlet {
 			myInsertStatement.setString(2, userID);
 			myInsertStatement.executeUpdate();
 			request.setAttribute("follows", "Dont Follow anymore");
-    	}else{
+    	}else if(request.getAttribute("follows").toString().equals("Follow")){
     		myConnection = myDB.getConnection("babble");
 			PreparedStatement myDeleteStatement = myConnection.prepareStatement("DELETE FROM follows WHERE follower = ? AND followee = ?");
 			myDeleteStatement.setString(1, initialUserID);
